Reorganize and document
